+++ src/layouts/OAuth.test.tsx
@@ -2,7 +2,8 @@ import * as React from 'react';
 import { shallow } from 'enzyme';
 import { isEmpty } from 'ramda';
 
-import { OAuthCallbackPage, splitIntoTwo, parseQueryParams } from 'src/layouts/OAuth';
+import { OAuthCallbackPage } from 'src/layouts/OAuth';
+import { splitIntoTwo, parseQueryParams } from 'src/utilities/queryParams';
 
 describe('layouts/OAuth', () => {
   const dispatch = jest.fn();
